fix widget bug

+++ App/Views/MatchView/KDAView.swift
@@ -11,14 +11,14 @@ struct KDAView: View {
     var kills: Int
     var deaths: Int
     var assists: Int
-    var size: Int
+    var size: CGFloat
     var body: some View {
         HStack(spacing: 0) {
             Text("\(kills)").bold()
             Text("/\(deaths)").lineLimit(1).foregroundColor(Color(.systemRed))
             Text("/\(assists)").lineLimit(1)
             Text(" (\(calculateKDA().rounded(toPlaces: 1).description))").bold().foregroundColor(Color(.systemGray))
-        }.frame(minWidth:45).font(.custom(fontString, size: CGFloat(size)))
+        }.frame(minWidth:45).font(.custom(fontString, size: size))
     }
     
     private func calculateKDA() -> Double {
